+++ src/main/java/net/frozenblock/wilderwild/worldgen/feature/configured/WWConfiguredFeatures.java
@@ -1966,7 +1966,8 @@ public final class WWConfiguredFeatures {
 						WWBlocks.YELLOW_HIBISCUS.defaultBlockState(),
 						WWBlocks.WHITE_HIBISCUS.defaultBlockState(),
 						WWBlocks.PINK_HIBISCUS.defaultBlockState(),
-						WWBlocks.PINK_HIBISCUS.defaultBlockState()
+						WWBlocks.PURPLE_HIBISCUS.defaultBlockState(),
+						WWBlocks.PURPLE_HIBISCUS.defaultBlockState()
 					)
 				)
 			)